get_last_line() from fba_launch_io.py to utils.py
