+++ py/fiberassign/fba_launch_io.py
@@ -55,7 +55,7 @@ import desimeter
 import fiberassign
 from fiberassign.scripts.assign import parse_assign, run_assign_full
 from fiberassign.assign import merge_results, minimal_target_columns
-from fiberassign.utils import Logger, assert_isoformat_utc, get_svn_version
+from fiberassign.utils import Logger, assert_isoformat_utc, get_svn_version, get_last_line
 
 # matplotlib
 import matplotlib.pyplot as plt
@@ -100,32 +100,6 @@ def get_latest_rundate(log=Logger.get(), step="", start=time()):
         rundate += "+00:00"
     log.info("{:.1f}s\t{}\tlatest rundate: {}".format(time() - start, step, rundate))
     return rundate
-
-
-def get_last_line(fn):
-    """
-    Return the last line of a text file.
-
-    Args:
-        fn: file name (string)
-
-    Returns:
-        last_line: (string)
-
-    Notes:
-        Fails if fn has one line only; we do not protect for that case,
-            as this function is intended to be used in get_program_latest_timestamp()
-            to read *ecsv ledgers, which will always have more than one line,
-            and we want the fastest function possible, to use in fiberassign on-the-fly.
-        Copied from https://stackoverflow.com/questions/46258499/how-to-read-the-last-line-of-a-file-in-python.
-    """
-    with open(fn, "rb") as f:
-        f.seek(-2, os.SEEK_END)
-        while f.read(1) != b"\n":
-            f.seek(-2, os.SEEK_CUR)
-        last_line = f.readline().decode().strip()
-    f.close()
-    return last_line
 
 
 def read_ecsv_keys(fn):
